Add the main views and templates for those views
Add answer, add, and edit views for answering, adding, and editing
questions, and a show_question view for showing a question when clicked
on. Also add templates for all of these views. In add, edit, and answer
views, just list all the questions for now with a to the show question
view displaying the question text.