Allow tags attribute when updating/adding a question, and make text boxes bigger
Before, when updating a question and adding a new question, only
question and answer attributes were allowed to be modified, change it
so that tags can be modified as well. Tags can only contain alphanumeric
characters and commas to separate tags. Also, make text boxes for
question and answer attributes bigger in adding/updating questions