On view questions view where questions are listed, add colours to represent the state a question is in relative to its review date
If questions are ready to be reviewed and are not overdue, make them
red. If they are overdue, make them black. If questions will be ready
soon, then make them orange and if they won't be ready soon, make them
green. Currently, the question's counter remains the same if greater
than 2x the review period and resets the counter to 0 if greater than
3x the review period. Make it so that it doesn't ever reset, only
remains the same if greater than 2x