Add support for logging in and out, and authentication to prevent users that aren't logged in from accessing any views (using django's login decorators)
Use Django's authentication support to create templates for login and
logout. Use Django's default views for these. When logging in, user is
redirected to home, and when logging out, or when trying to access a page
without being logged in, user is redirected to login screen. Note: There
is still no support for registering/signing up a user and different users
having access to their own questions