Add a sign up page for creating a new user
Also, when a user that is already signed tries to go to the login page or
signup page, redirect them to the home page. This is because they
shouldn't be able to access those pages unless they are logged out. If
they try to access the logout page, the django logout view logs them out
so that doesn't pose a problem