When a user is logged in, limit their account to their own questions. They can only answer, view, delete, etc. their own questions and can not interact with other user's questions at all. Basically, every user has their own "account", or at least the equivalent of it
Before this commit, all users have access to view and answer all the
questions in the database. Update it so that the user can only view and
answer their own questions, and if they try to visit a url of another
user's question, a 404 error page will be displayed. Also, fix colours
in edit view since the legend doesn't match the question link colours